# contracts/openapi/topsdojob-v3-local.yaml
openapi: 3.1.0
info:
  title: Tops do Job V3 Local API
  version: 0.0.1
  description: "Contrato minimo local do skeleton, camada transversal, API publica de leitura e funil local Anuncie gratis. A API segue fail-closed: endpoints nao descritos em /api/** nao devem ficar publicos por acidente."
servers:
  - url: http://localhost:8080
    description: backend local
paths:
...
  /api/admin/anuncios/{id}:
    get:
      summary: Detalhe administrativo sanitizado de anuncio
      description: Retorna os dados comerciais necessarios a moderacao, proprietario mascarado e revisao aberta, somente para ADMIN ou MODERADOR com ANUNCIO_LER.
      operationId: getAdminAnuncioDetalhe
      security:
        - AdminSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: detalhe sanitizado de anuncio
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAnuncioDetalhe"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/admin/premium/anuncios/{id}:
    get:
      summary: Status Premium read-only de anuncio
      description: Retorna calculo local sanitizado de Premium/beneficios para ADMIN ou MODERADOR. Nao compra, nao ativa, nao cobra, nao chama Pix/Efi, nao retorna valor pago, saldo, credito real, payload de pagamento, telefone bruto, storage ou documento privado.
      operationId: getAdminPremiumAnuncioStatus
      security:
        - AdminSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: status Premium sanitizado e somente leitura
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPremiumAnuncioStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/admin/premium/anuncios/{id}/beneficios:
    get:
      summary: Beneficios Premium read-only de anuncio
      description: Lista beneficios calculados de forma sanitizada. Nao retorna custo, preco, saldo, origem financeira detalhada, payload, idempotency key ou dado sensivel.
      operationId: getAdminPremiumBeneficiosAnuncio
      security:
        - AdminSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: beneficios sanitizados do anuncio
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminBeneficioAnuncio"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/admin/premium/consistencia:
    get:
      summary: Consistencia Premium read-only
      description: Retorna inconsistencias locais calculadas em dados sinteticos, incluindo expiracao conjunta por grupo/pacote/campanha. Nao altera banco, nao expira beneficio de verdade, nao agenda job e nao executa acao financeira.
      operationId: getAdminPremiumConsistencia
      security:
        - AdminSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: relatorio de consistencia Premium
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPremiumConsistenciaResumo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/admin/premium/vencendo:
    get:
      summary: Beneficios Premium vencendo
      description: Retorna beneficios vencendo na janela local de sete dias. Somente leitura, sem alerta externo, sem envio, sem checkout, sem Pix/Efi e sem alteracao de credito.
      operationId: getAdminPremiumVencendo
      security:
        - AdminSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: beneficios vencendo em janela local
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPremiumVencendoResumo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/admin/premium/catalogo:
    get:
      summary: Lista catalogo Premium administrativo
      operationId: getAdminPremiumCatalogo
      security:
        - AdminSessionCookie: []
      responses:
        "200":
          description: beneficios e opcoes vigentes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PremiumCatalogo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/premium/catalogo/{id}:
    put:
      summary: Atualiza descricao, ordem, status, custos e duracoes Premium
      operationId: putAdminPremiumCatalogo
      security:
        - AdminSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminPremiumCatalogoUpdateRequest"
      responses:
        "200":
          description: catalogo atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PremiumCatalogo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/admin/premium/ativacoes:
    get:
      summary: Lista ativacoes Premium de um usuario
      operationId: getAdminPremiumAtivacoes
      security:
        - AdminSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: usuarioId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: ativacoes do usuario
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminPremiumAtivacao"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/premium/ativacoes/{id}/cancelar:
    post:
      summary: Cancela ativacao e estorna credito quando permitido
      operationId: postAdminPremiumCancelar
      security:
        - AdminSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [motivo]
              properties:
                motivo:
                  type: string
                  minLength: 5
                  maxLength: 500
      responses:
        "200":
          description: ativacao cancelada ou repeticao idempotente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPremiumAtivacaoOperacao"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    PublicSessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Sessao publica emitida pelo login; cookie HttpOnly, Secure em HML e SameSite Lax.
    AdminSessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Sessao administrativa local emitida por `POST /api/admin/auth/login`; cliente deve usar credentials include e o corpo da resposta nunca retorna token ou cookie.
  headers:
    RequestId:
      description: Identificador opaco de correlação da requisição.
      schema:
        type: string
        example: 11111111-1111-4111-8111-111111111111
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Chave opaca para impedir repeticao da mesma operacao mutavel.
      schema:
        type: string
        minLength: 1
        maxLength: 160
        pattern: "^[A-Za-z0-9._:-]+$"
...
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      description: Identificador opaco enviado pelo cliente para correlação.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: "^[A-Za-z0-9._:-]{8,128}$"
...
  responses:
    BadRequest:
      description: Parametro invalido ou rota publica proibida.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    NotFound:
      description: Recurso publico nao encontrado.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    Unauthorized:
      description: Autenticacao necessaria ou credenciais invalidas.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    Forbidden:
      description: Usuario autenticado sem permissao suficiente.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    Conflict:
      description: Estado atual impede a transicao solicitada.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
...
    InternalError:
      description: Falha inesperada sem detalhes internos.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
  schemas:
...
    AdminLocalizacaoSanitizada:
      type: object
      additionalProperties: false
      properties:
        uf:
          type:
            - string
            - "null"
        cidade:
          type:
            - string
            - "null"
        bairro:
          type:
            - string
            - "null"
    AdminAnuncianteResumo:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        nome:
          type:
            - string
            - "null"
        emailMascarado:
          type:
            - string
            - "null"
          description: Identificador visual mascarado; nunca retorna o e-mail integral.
        status:
          type:
            - string
            - "null"
    AdminRevisaoAberta:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        tipo:
          type: string
        status:
          type: string
          enum: [ABERTA, EM_ANALISE]
        criadoEm:
          type: string
          format: date-time
    AdminAnuncioListaItem:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        slug:
          type:
            - string
            - "null"
        titulo:
          type:
            - string
            - "null"
        status:
          type:
            - string
            - "null"
        statusModeracao:
          type:
            - string
            - "null"
        localizacao:
          anyOf:
            - $ref: "#/components/schemas/AdminLocalizacaoSanitizada"
            - type: "null"
        criadoEm:
          type:
            - string
            - "null"
          format: date-time
        atualizadoEm:
          type:
            - string
            - "null"
          format: date-time
        publicadoEm:
          type:
            - string
            - "null"
          format: date-time
        midiasTotal:
          type:
            - integer
            - "null"
          format: int64
        revisoesTotal:
          type:
            - integer
            - "null"
          format: int64
        contatoConfigurado:
          type: boolean
          description: Indicador agregado; nao retorna telefone nem WhatsApp.
        documentoPendente:
          type: boolean
          description: Indicador agregado; nao retorna documento privado.
        comercialLimitado:
          type: boolean
        anunciante:
          anyOf:
            - $ref: "#/components/schemas/AdminAnuncianteResumo"
            - type: "null"
        revisaoAberta:
          anyOf:
            - $ref: "#/components/schemas/AdminRevisaoAberta"
            - type: "null"
    AdminAnuncioDetalhe:
      allOf:
        - $ref: "#/components/schemas/AdminAnuncioListaItem"
        - type: object
          additionalProperties: false
          properties:
            descricaoResumo:
              type:
                - string
                - "null"
              description: Texto sanitizado e mascarado; pode ser nulo quando ausente na origem.
            descricao:
              type:
                - string
                - "null"
              description: Conteudo comercial integral para moderacao; pode ser nulo quando ausente.
            categoria:
              type:
                - string
                - "null"
            ultimaPublicacaoEm:
              type:
                - string
                - "null"
              format: date-time
            precoInformado:
              type: boolean
            preco:
              type:
                - number
                - "null"
              format: double
              description: Preco informado no anuncio; pode ser nulo quando nao informado.
            whatsapp:
              type:
                - string
                - "null"
              description: Contato do proprio anuncio para analise; pode ser nulo quando nao informado.
            locaisAtendimento:
              type: array
              items:
                type: string
            servicos:
              type: array
              items:
                type: string
    AdminBeneficioAnuncio:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        beneficioCodigo:
          type:
            - string
            - "null"
          description: Codigo operacional do beneficio. Nao carrega valor pago.
        beneficioNome:
          type:
            - string
            - "null"
        escopo:
          type:
            - string
            - "null"
          enum:
            - ANUNCIO
            - USUARIO
            - MIDIA
            - RELATORIO
            - null
        statusOriginal:
          type:
            - string
            - "null"
        statusCalculado:
          type:
            - string
            - "null"
          enum:
            - ATIVO
            - EXPIRADO
            - VENCENDO
            - PENDENTE
            - INATIVO
            - INCONSISTENTE
            - null
        inicioEm:
          type:
            - string
            - "null"
          format: date-time
        fimEm:
          type:
            - string
            - "null"
          format: date-time
        venceEmBreve:
          type: boolean
        grupoVinculado:
          type: boolean
        grupoId:
          type:
            - string
            - "null"
          format: uuid
        grupoTipo:
          type:
            - string
            - "null"
        grupoStatus:
          type:
            - string
            - "null"
        grupoFimEm:
          type:
            - string
            - "null"
          format: date-time
        codigosConsistencia:
          type: array
          items:
            type: string
        inconsistente:
          type: boolean
        somenteLeitura:
          type: boolean
          const: false
    AdminPremiumAnuncioStatus:
      type: object
      additionalProperties: false
      properties:
        anuncioId:
          type: string
          format: uuid
        slug:
          type:
            - string
            - "null"
        titulo:
          type:
            - string
            - "null"
          description: Titulo sanitizado.
        premiumAtivo:
          type: boolean
        destaqueAtivo:
          type: boolean
        topoAtivo:
          type: boolean
        possuiStories:
          type: boolean
        possuiMidiaExtra:
          type: boolean
        beneficiosAtivos:
          type: integer
        beneficiosExpirados:
          type: integer
        beneficiosVencendo:
          type: integer
        inconsistenciasTotal:
          type: integer
        codigosConsistencia:
          type: array
          items:
            type: string
        calculadoEm:
          type:
            - string
            - "null"
          format: date-time
        somenteLeitura:
          type: boolean
          const: true
        compraOuAtivacaoRealDisponivel:
          type: boolean
          const: false
        acoesFinanceirasDisponiveis:
          type: boolean
          const: false
        gratuitoLimitadoPorContato:
          type: boolean
          const: false
        valorPago:
          not: {}
          description: Campo proibido.
        saldoCredito:
          not: {}
          description: Campo proibido.
        payloadPagamento:
          not: {}
          description: Campo proibido.
    AdminPremiumConsistenciaItem:
      type: object
      additionalProperties: false
      properties:
        anuncioId:
          type:
            - string
            - "null"
          format: uuid
        slug:
          type:
            - string
            - "null"
        ativacaoId:
          type:
            - string
            - "null"
          format: uuid
        beneficioCodigo:
          type:
            - string
            - "null"
        codigo:
          type: string
          enum:
            - GRUPO_EXPIRADO_COM_BENEFICIO_ATIVO
            - BENEFICIO_EXPIRADO_ANTES_DO_GRUPO
            - GRUPO_SEM_BENEFICIOS
            - DATA_INVALIDA
            - ORIGEM_DESCONHECIDA
        severidade:
          type: string
        mensagem:
          type: string
        statusCalculado:
          type:
            - string
            - "null"
        fimEm:
          type:
            - string
            - "null"
          format: date-time
        grupoId:
          type:
            - string
            - "null"
          format: uuid
        grupoStatus:
          type:
            - string
            - "null"
        grupoFimEm:
          type:
            - string
            - "null"
          format: date-time
        detectadoEm:
          type:
            - string
            - "null"
          format: date-time
    AdminPremiumConsistenciaResumo:
      type: object
      additionalProperties: false
      properties:
        itens:
          type: array
          items:
            $ref: "#/components/schemas/AdminPremiumConsistenciaItem"
        total:
          type: integer
        calculadoEm:
          type:
            - string
            - "null"
          format: date-time
        somenteLeitura:
          type: boolean
          const: true
    AdminPremiumVencendoItem:
      type: object
      additionalProperties: false
      properties:
        anuncioId:
          type:
            - string
            - "null"
          format: uuid
        slug:
          type:
            - string
            - "null"
        ativacaoId:
          type: string
          format: uuid
        beneficioCodigo:
          type:
            - string
            - "null"
        statusCalculado:
          type:
            - string
            - "null"
        fimEm:
          type:
            - string
            - "null"
          format: date-time
        diasRestantes:
          type: integer
        grupoVinculado:
          type: boolean
        somenteLeitura:
          type: boolean
          const: true
    AdminPremiumVencendoResumo:
      type: object
      additionalProperties: false
      properties:
        itens:
          type: array
          items:
            $ref: "#/components/schemas/AdminPremiumVencendoItem"
        total:
          type: integer
        janelaDias:
          type: integer
          const: 7
        calculadoEm:
          type:
            - string
            - "null"
          format: date-time
        somenteLeitura:
          type: boolean
          const: true
...
    ApiErrorResponse:
      type: object
      additionalProperties: false
      required:
        - timestamp
        - status
        - error
        - code
        - message
        - path
        - requestId
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 500
        error:
          type: string
          example: Internal Server Error
        code:
          $ref: "#/components/schemas/ApiErrorCode"
        message:
          type: string
          example: Erro interno inesperado.
        path:
          type: string
          example: /api/health
        requestId:
          type: string
          example: 11111111-1111-4111-8111-111111111111
    ApiErrorCode:
      type: string
      enum:
        - BAD_REQUEST
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - UNPROCESSABLE_ENTITY
        - BAD_GATEWAY
        - SERVICE_UNAVAILABLE
        - INTERNAL_ERROR
...
    PremiumOpcao:
      type: object
      additionalProperties: false
      required: [id, duracaoDias, custoCreditos, ativo, ordemExibicao]
      properties:
        id: { type: string, format: uuid }
        duracaoDias: { type: integer, enum: [1, 7, 14, 30] }
        custoCreditos: { type: integer, minimum: 0 }
        ativo: { type: boolean }
        ordemExibicao: { type: integer, minimum: 0 }
    PremiumCatalogo:
      type: object
      additionalProperties: false
      required: [id, codigo, nome, descricao, escopo, afetaRanking, ativo, ordemExibicao, opcoes]
      properties:
        id: { type: string, format: uuid }
        codigo: { type: string }
        nome: { type: string }
        descricao: { type: string }
        escopo: { type: string }
        afetaRanking: { type: boolean }
        ativo: { type: boolean }
        ordemExibicao: { type: integer, minimum: 0 }
        opcoes:
          type: array
          items:
            $ref: "#/components/schemas/PremiumOpcao"
...
    AdminPremiumAtivacao:
      type: object
      additionalProperties: false
      required: [id, usuarioId, anuncioId, status, custoCreditos, inicioEm, fimEm]
      properties:
        id: { type: string, format: uuid }
        usuarioId: { type: string, format: uuid }
        anuncioId: { type: string, format: uuid }
        beneficioCodigo: { type: [string, "null"] }
        beneficioNome: { type: [string, "null"] }
        origem: { type: [string, "null"] }
        status: { type: string }
        custoCreditos: { type: integer, minimum: 0 }
        inicioEm: { type: string, format: date-time }
        fimEm: { type: string, format: date-time }
    AdminPremiumAtivacaoOperacao:
      type: object
      additionalProperties: false
      required: [id, usuarioId, anuncioId, beneficioId, status, creditosEstornados, idempotente]
      properties:
        id: { type: string, format: uuid }
        usuarioId: { type: string, format: uuid }
        anuncioId: { type: string, format: uuid }
        beneficioId: { type: string, format: uuid }
        status: { type: string }
        creditosEstornados: { type: integer, minimum: 0 }
        fimEm: { type: [string, "null"], format: date-time }
        idempotente: { type: boolean }
    AdminPremiumCatalogoUpdateRequest:
      type: object
      additionalProperties: false
      required: [nome, descricao, ativo, ordemExibicao, opcoes]
      properties:
        nome: { type: string, minLength: 2, maxLength: 120 }
        descricao: { type: string, minLength: 5, maxLength: 500 }
        ativo: { type: boolean }
        ordemExibicao: { type: integer, minimum: 0 }
        opcoes:
          type: array
          minItems: 1
          items:
            type: object
            additionalProperties: false
            required: [duracaoDias, custoCreditos, ativo, ordemExibicao]
            properties:
              duracaoDias: { type: integer, enum: [1, 7, 14, 30] }
              custoCreditos: { type: integer, minimum: 0 }
              ativo: { type: boolean }
              ordemExibicao: { type: integer, minimum: 0 }
...